refactor(cli): stop bundling server-only services/utils code into the CLI (#660)
The b4m CLI imported the @bike4mind/services barrel and the @bike4mind/utils
barrel, dragging server-only code it never runs (user registration, image
moderation, xlsx export, image downscaling) plus their heavy npm deps into its
bundle. The #637 build guard had to declare five of those leaked deps to keep
the CLI installable.

Draw the missing import boundaries so the CLI graph can't reach them:

- services: extract the pure tool/MCP generators (generateTools,
  generateMcpTools, generateMcpToolsFromCache) into an implementation-free
  toolGenerators module. tools/index re-exports them so the server barrel API is
  unchanged. generateTools no longer defaults its `tools` arg to b4mTools (that
  default was the edge that pulled the whole tool graph); the one caller relying
  on it now passes b4mTools explicitly.
- services: cliTools (the existing CLI-isolated entry) now also exposes the
  small subset of shared tools the CLI enables plus the generators, so the CLI
  imports @bike4mind/services/llm/tools/cliTools instead of the barrel.
- utils: move imageModeration (Rekognition + jimp) and the jimp image-resize
  helper off the barrel into @bike4mind/utils/imageModeration and
  @bike4mind/utils/imageResize subpaths. utils' dist barrel is a single bundled
  module, so anything on it is reachable from every importer; the guard follows
  the dynamic import() edges regardless of tree-shaking. processFabFilesServer
  now takes the resize helper as an injected dependency so llm/utils carries no
  jimp reference.

Result: @aws-sdk/client-rekognition, jimp, write-excel-file, and
disposable-email-domains are removed from packages/cli/package.json and the
externals guard stays green. CLI bundle shrinks 2.42 MB -> 2.23 MB (-7.9%);
baseline.json refreshed.

Closes #660
