+++ apps/client/server/queueHandlers/imageEdit.ts
@@ -14,7 +14,8 @@ import {
   usageEventRepository,
   userRepository,
 } from '@bike4mind/database';
-import { RekognitionImageModerationService, SQSService } from '@bike4mind/utils';
+import { SQSService } from '@bike4mind/utils';
+import { RekognitionImageModerationService } from '@bike4mind/utils/imageModeration';
 import { Logger } from '@bike4mind/observability';
 import { logEvent } from '@server/utils/analyticsLog';
 import { dispatchWithLogger } from '@server/queueHandlers/utils';
