+++ b4m-core/services/src/llm/ImageGeneration.ts
@@ -57,8 +57,8 @@ import {
   ImageEditResponse,
   BaseStorage,
   getSettingsByNames,
-  ImageModerationService,
 } from '@bike4mind/utils';
+import type { ImageModerationService } from '@bike4mind/utils/imageModeration';
 import { getAvailableModels } from '@bike4mind/llm-adapters';
 import { Logger } from '@bike4mind/observability';
 import { MongoAbility } from '@casl/ability';
