+++ packages/cli/package.json
@@ -54,7 +54,6 @@
     "@aws-sdk/client-bedrock-runtime": "^3.1080.0",
     "@aws-sdk/client-cloudwatch": "^3.1080.0",
     "@aws-sdk/client-lambda": "^3.1080.0",
-    "@aws-sdk/client-rekognition": "^3.1080.0",
     "@aws-sdk/client-s3": "^3.1080.0",
     "@aws-sdk/client-sqs": "^3.1080.0",
     "@aws-sdk/client-transcribe": "^3.1080.0",
@@ -78,7 +77,6 @@
     "csv-parse": "^6.2.1",
     "dayjs": "^1.11.21",
     "diff": "^9.0.0",
-    "disposable-email-domains": "^1.0.62",
     "dotenv": "^17.4.2",
     "eventsource-parser": "^3.1.0",
     "fdir": "^6.5.0",
@@ -92,7 +90,6 @@
     "ink-select-input": "^6.2.0",
     "ink-spinner": "^5.0.0",
     "ink-text-input": "^6.0.0",
-    "jimp": "^1.6.1",
     "jsonwebtoken": "^9.0.3",
     "jszip": "^3.10.1",
     "lodash": "^4.18.1",
@@ -120,7 +117,6 @@
     "uuid": "^13.0.2",
     "voyageai": "^0.4.0",
     "web-tree-sitter": "0.25.10",
-    "write-excel-file": "^4.1.1",
     "ws": "^8.21.0",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz",
     "yargs": "^18.0.0",
